backend/requirements.txt
Django==2.2.19
pytest==6.2.4
pytest-django==4.4.0
python-dotenv==0.20.0
django-extensions==3.1.5
ipython==8.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
django-filter==21.1